refactor: update imports and enhance sorting and status handling in prediction market API
